Biblioteca/requirements.txt
opencv-python==4.8.0.74
Pillow==10.0.0
progress==1.6
twine==4.0.2
